+++ app/src/main/java/com/example/myapplication/ui/dashboard/Json.java
@@ -1,0 +1,5 @@
+package com.example.myapplication.ui.dashboard;
+
+public interface Json {
+    String connect();
+}
