+++ app/src/main/java/com/example/myapplication/ui/dashboard/watchHistory.java
@@ -1,8 +1,6 @@
 package com.example.myapplication.ui.dashboard;
 
-import android.annotation.SuppressLint;
 import android.content.Intent;
-import android.os.AsyncTask;
 import android.os.Bundle;
 import android.util.Log;
 import android.widget.ListAdapter;
@@ -13,15 +11,6 @@ import androidx.appcompat.app.AppCompatActivity;
 
 import com.example.myapplication.R;
 
-import org.json.JSONArray;
-import org.json.JSONException;
-import org.json.JSONObject;
-
-import java.io.IOException;
-import java.io.InputStream;
-import java.io.InputStreamReader;
-import java.net.HttpURLConnection;
-import java.net.URL;
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
@@ -41,14 +30,15 @@ public class watchHistory extends AppCompatActivity {
         Intent myIntent = getIntent();
         dateString = myIntent.getStringExtra("date");
 
-
         lv = findViewById(R.id.Listview);
         resultList = myIntent.getStringArrayListExtra("resultList");
         List<HashMap<String, String>> list = new ArrayList<>();
         for(String result : resultList) {
             HashMap<String, String> item = new HashMap<>();
             item.put("result", result);
+            list.add(item);
         }
+
         //Displaying the results
         ListAdapter adapter = new SimpleAdapter(
                 watchHistory.this,
